Remove unnecessary MockitoExtension from Spec11PipelineTest
This is kind of a shot in the dark here, but this is one of the obvious
differences between this test class (which frequently experiences flakes) and
the other pipeline test classes which do not.

I also explicitly set the runner in the Spec11PipelineOptions, which seems
possibly related to the NPEs we're seeing in the flakes when it fails to get the
ThreadFactory. This is the only test that explicitly sets these options, so it's
not clear if the usual default to CrashingRunner works if the options are
overridden.