Fix flaky Spec11PipelineTest
It appears when the test fails the call to isAppEngine() returns true. It
is still not clear where that comes from, but it is likely due to the
DatastoreEntityExtension trying to setup an App Engine environment for
Ofy. All Ofy needs is a project ID so we explicitly clear the system
proporty for ApiProxy in the hope that isAppEngine() will now always
return false.