Add methods to return subtypes of HistoryEntry when querying
This is useful when we expect a specfic subtype in the return value so
that we can set the parent resource (e. g. DomainContent for
DomainHistory) on it, or when a specific subtype is needed from the call
site.

Also fixed some generic return values. It is always to return
<HistoryEntry> than a wildcard <? extends HistoryEntry> because for immutable
collections <? extends HistoryEntry> is no different than <HistoryEntry> as
return value -- you can only get a HistoryEntry from it. The wildcard
means that even if you are indeed getting a <DomainHistory> from
querying, the call site has no compile time knowledge of it and can only
assume it is a HistoryEntry.