Fix the runtime and correctness of replaying commit logs to SQL
- We limit the search to 30 minutes after the last checkpoint to narrow
the scope.

- We now use a more intelligent prefix to narrow the listObjects search
space in GCS. Otherwise, we're returning >30k objects which can take
roughly 50 seconds. This results in a listObjects time of 1-3 seconds.

- When retrieving the object info for each file in GCS, the threads
should be created using the standard Java executor service. If we use
the request thread factory (as it did before), it attempts to create
background threads (threads with no ties to the request in question).
For starters, we don't mind the threads having a connection to the
request in question, but more importantly AppEngine limits us to 10
background threads per instance
(https://cloud.google.com/appengine/docs/standard/java/runtime#threads)
and some of these might be taken up by other requests / services. As a
result, much of the time the request to GCS never actually happens.