Create and use a ReadOnlyTransactionManager for SQL migration
We create two additional read-only transaction managers, one backed by
each existing transaction maanger. That way, when we are in the
read-only stage(s) of the migration we can return the appropriate
read-only transaction manager while still allowing the raw ofyTm() and
jpaTm() calls to delegate to the read-and-write transaction managers.