Fix hanging threads in GcsDiffFileLister
When retrieving the object info for each file in GCS, the threads
should be created using the standard Java executor service. Using the
request thread factory is generally a good thing, however AppEngine
restricts the creation of threads using that factory -- if you attempt
to create threads in a non-request thread, that request fails. This
failure, however, is manifested by a hanging rather than an actual
error, meaning we never actually find out what went wrong.

Because we don't exactly control which threads the executor is created
and run on, we should use the standard Java thread management rather
than the request thread factory to avoid this hanging.