Skip synthetic history entries for resources that don't need them
The reason for creating synthetic history entries is so that we can
guarantee that each EppResource has at least one corresponding *History
object that contains the resource at that point in time. If we have
history objects in SQL that contain a particular resource already, there
is no need to create a synthetic history entry for that resource.