Retry datastore transactions during replay
It looks like the datastore failures we've been seeing may very well have been
due to transient failures on the part of the test datastore.  Wrap the replay
transactions in a 5x retrier to remedy.

Note that this problem would not affect the actual replay, nor does this
solution: the actual replay uses a cron job and if a specific transaction
fails it will be retried on the next cycle (though we may consider adding
retry for that, if it's not already present).