Remove Ofy support from Cursor
Cursor was originally envisioned to support arbitrary ImmutableObject
scopes. However, in practice only the Registry scope is used. The SQL
representation of Cursor assumes that and the schema uses a composite ID
with a string column for the primary key of the scope object. Without a
schema migration to persist the VKey of the scope, we cannot support any
ImmutableObject other than those with a primitive string primary key.

Given the complexity involved and the limited use case, the scope is now
explicitly limited to Registry only.

Also made some code quality improvement based on IntelliJ suggestions on
modified files.