Remove ofy support from BillingEvent
This PR turns out to be more massive than I would have liked but it
deals with all billing event related stuff, which are more or link all
intertwined:

* Remove all billing events as Ofy entities.
* Add a temporary annotation to allow BillingEvent's ID to be
  auto-allocated by ofy while not lacking the Ofy @Id annotation.
* Remove Modification, which is only used in ofy.
* Remove BillingVKey, as we do not need to store the ofy key parent
   information anymore. The VKey for a billing event now just contain
   its primary key, and can be converted by VKeyConverter.
* Remove BigQuery related code in the billing pipeline.

Note that after BillingVKey is removed, several columns in
BillingCancellation are no longer needed. The change to database schema
will be handled in https://github.com/google/nomulus/pull/1721 after
this PR is deployed to production.