Fix GAE build under Gradle 7
For some reason after the upgrade to Gradle, the core.jar file is no
longer included in the generated WAR, even though the deploy_jar
configuration is specified as a dependency.

I could not figure out a way to tweak the configuration dependency to
have core.jar pulled into the .war, so I decided to just explicitly pick
it from its known location.

Also upgraded to the latest GAE Gradle plugin version.

TESTED=deployed to alpha and verified that the instances can start.